changed package Name and App Name
