+++ app/src/main/java/com/nanodegree/velaphi/popularmovies/models/ReviewResponse.java
@@ -1,0 +1,65 @@
+package com.nanodegree.velaphi.popularmovies.models;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+import java.util.List;
+
+public class ReviewResponse {
+
+    @SerializedName("id")
+    @Expose
+    private Integer id;
+    @SerializedName("page")
+    @Expose
+    private Integer page;
+    @SerializedName("results")
+    @Expose
+    private List<Review> reviews = null;
+    @SerializedName("total_pages")
+    @Expose
+    private Integer totalPages;
+    @SerializedName("total_results")
+    @Expose
+    private Integer totalResults;
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public Integer getPage() {
+        return page;
+    }
+
+    public void setPage(Integer page) {
+        this.page = page;
+    }
+
+    public List<Review> getReviews() {
+        return reviews;
+    }
+
+    public void setResults(List<Review> reviews) {
+        this.reviews = reviews;
+    }
+
+    public Integer getTotalPages() {
+        return totalPages;
+    }
+
+    public void setTotalPages(Integer totalPages) {
+        this.totalPages = totalPages;
+    }
+
+    public Integer getTotalResults() {
+        return totalResults;
+    }
+
+    public void setTotalResults(Integer totalResults) {
+        this.totalResults = totalResults;
+    }
+}
