+++ app/src/main/java/com/nanodegree/velaphi/popularmovies/repository/PopularMoviesRepositoryIml.java
@@ -1,0 +1,68 @@
+package com.nanodegree.velaphi.popularmovies.repository;
+
+import android.arch.lifecycle.LiveData;
+import android.support.annotation.NonNull;
+
+import com.nanodegree.velaphi.popularmovies.database.FavoriteMovieDatabase;
+import com.nanodegree.velaphi.popularmovies.models.Movie;
+import com.nanodegree.velaphi.popularmovies.models.PopularMoviesResponse;
+import com.nanodegree.velaphi.popularmovies.remote.AppClient;
+import com.nanodegree.velaphi.popularmovies.remote.RequestInterface;
+
+import java.util.List;
+
+import javax.inject.Inject;
+
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
+
+public class PopularMoviesRepositoryIml implements PopularMoviesRepository {
+
+    @Inject
+    FavoriteMovieDatabase favoriteMovieDatabase;
+
+    public PopularMoviesRepositoryIml(FavoriteMovieDatabase favoriteMovieDatabase) {
+        this.favoriteMovieDatabase = favoriteMovieDatabase;
+    }
+
+    @Override
+    public void getTopRatedMovies(@NonNull final RepositoryCallback repositoryCallback) {
+        RequestInterface requestInterface = AppClient.getApi();
+
+        requestInterface.getTopRated().enqueue(new Callback<PopularMoviesResponse>() {
+            @Override
+            public void onResponse(Call<PopularMoviesResponse> call, Response<PopularMoviesResponse> response) {
+                repositoryCallback.onSuccess(response.body());
+            }
+
+            @Override
+            public void onFailure(Call<PopularMoviesResponse> call, Throwable t) {
+                repositoryCallback.onFailure(t);
+            }
+        });
+    }
+
+    @Override
+    public void getPopularMovies(@NonNull final RepositoryCallback repositoryCallback) {
+        RequestInterface requestInterface = AppClient.getApi();
+
+        requestInterface.getPopular().enqueue(new Callback<PopularMoviesResponse>() {
+            @Override
+            public void onResponse(Call<PopularMoviesResponse> call, Response<PopularMoviesResponse> response) {
+                repositoryCallback.onSuccess(response.body());
+            }
+
+            @Override
+            public void onFailure(Call<PopularMoviesResponse> call, Throwable t) {
+                repositoryCallback.onFailure(t);
+            }
+        });
+    }
+
+
+    @Override
+    public LiveData<List<Movie>>getFavouriteMovies() {
+        return favoriteMovieDatabase.getFavoriteMovieDao().getFavoriteMovies();
+    }
+}
