--- app/src/main/java/com/nanodegree/velaphi/popularmoviesstage1/models/Movie.java
@@ -1,306 +1,0 @@
-package com.nanodegree.velaphi.popularmoviesstage1.models;
-
-import android.arch.persistence.room.ColumnInfo;
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.Ignore;
-import android.arch.persistence.room.PrimaryKey;
-import android.os.Parcel;
-import android.os.Parcelable;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
-
-@Entity(tableName = "favorite_movies")
-public class Movie implements Parcelable {
-
-    @ColumnInfo
-    @SerializedName("vote_count")
-    @Expose
-    private Integer voteCount;
-
-    @PrimaryKey
-    @SerializedName("id")
-    @Expose
-    private Integer id;
-
-    @ColumnInfo
-    @SerializedName("video")
-    @Expose
-    private Boolean video;
-
-    @ColumnInfo(name = "vote_average")
-    @SerializedName("vote_average")
-    @Expose
-    private Double voteAverage;
-
-    @ColumnInfo
-    @SerializedName("title")
-    @Expose
-    private String title;
-
-    @ColumnInfo
-    @SerializedName("popularity")
-    @Expose
-    private Double popularity;
-
-    @ColumnInfo(name = "poster_path" )
-    @SerializedName("poster_path")
-    @Expose
-    private String posterPath;
-
-    @ColumnInfo(name = "original_language")
-    @SerializedName("original_language")
-    @Expose
-    private String originalLanguage;
-
-    @ColumnInfo(name = "original_title")
-    @SerializedName("original_title")
-    @Expose
-    private String originalTitle;
-
-    @Ignore
-    @SerializedName("genre_ids")
-    @Expose
-    private List<Integer> genreIds = null;
-
-    @ColumnInfo(name = "backdrop_path")
-    @SerializedName("backdrop_path")
-    @Expose
-    private String backdropPath;
-
-    @SerializedName("adult")
-    @Expose
-    private Boolean adult;
-
-    @ColumnInfo
-    @SerializedName("overview")
-    @Expose
-    private String overview;
-
-    @ColumnInfo(name = "release_date")
-    @SerializedName("release_date")
-    @Expose
-    private String releaseDate;
-
-    public Movie(Integer voteCount, Integer id, Boolean video, Double voteAverage,
-                 String title, Double popularity, String posterPath, String originalLanguage,
-                 String originalTitle, String backdropPath, Boolean adult, String overview,
-                 String releaseDate) {
-        this.voteCount = voteCount;
-        this.id = id;
-        this.video = video;
-        this.voteAverage = voteAverage;
-        this.title = title;
-        this.popularity = popularity;
-        this.posterPath = posterPath;
-        this.originalLanguage = originalLanguage;
-        this.originalTitle = originalTitle;
-        this.backdropPath = backdropPath;
-        this.adult = adult;
-        this.overview = overview;
-        this.releaseDate = releaseDate;
-    }
-
-    protected Movie(Parcel in) {
-        if (in.readByte() == 0) {
-            voteCount = null;
-        } else {
-            voteCount = in.readInt();
-        }
-        if (in.readByte() == 0) {
-            id = null;
-        } else {
-            id = in.readInt();
-        }
-        byte tmpVideo = in.readByte();
-        video = tmpVideo == 0 ? null : tmpVideo == 1;
-        if (in.readByte() == 0) {
-            voteAverage = null;
-        } else {
-            voteAverage = in.readDouble();
-        }
-        title = in.readString();
-        if (in.readByte() == 0) {
-            popularity = null;
-        } else {
-            popularity = in.readDouble();
-        }
-        posterPath = in.readString();
-        originalLanguage = in.readString();
-        originalTitle = in.readString();
-        backdropPath = in.readString();
-        byte tmpAdult = in.readByte();
-        adult = tmpAdult == 0 ? null : tmpAdult == 1;
-        overview = in.readString();
-        releaseDate = in.readString();
-    }
-
-    public static final Creator<Movie> CREATOR = new Creator<Movie>() {
-        @Override
-        public Movie createFromParcel(Parcel in) {
-            return new Movie(in);
-        }
-
-        @Override
-        public Movie[] newArray(int size) {
-            return new Movie[size];
-        }
-    };
-
-    public Integer getVoteCount() {
-        return voteCount;
-    }
-
-    public void setVoteCount(Integer voteCount) {
-        this.voteCount = voteCount;
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public Boolean getVideo() {
-        return video;
-    }
-
-    public void setVideo(Boolean video) {
-        this.video = video;
-    }
-
-    public Double getVoteAverage() {
-        return voteAverage;
-    }
-
-    public void setVoteAverage(Double voteAverage) {
-        this.voteAverage = voteAverage;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public Double getPopularity() {
-        return popularity;
-    }
-
-    public void setPopularity(Double popularity) {
-        this.popularity = popularity;
-    }
-
-    public String getPosterPath() {
-        return posterPath;
-    }
-
-    public void setPosterPath(String posterPath) {
-        this.posterPath = posterPath;
-    }
-
-    public String getOriginalLanguage() {
-        return originalLanguage;
-    }
-
-    public void setOriginalLanguage(String originalLanguage) {
-        this.originalLanguage = originalLanguage;
-    }
-
-    public String getOriginalTitle() {
-        return originalTitle;
-    }
-
-    public void setOriginalTitle(String originalTitle) {
-        this.originalTitle = originalTitle;
-    }
-
-    public List<Integer> getGenreIds() {
-        return genreIds;
-    }
-
-    public void setGenreIds(List<Integer> genreIds) {
-        this.genreIds = genreIds;
-    }
-
-    public String getBackdropPath() {
-        return backdropPath;
-    }
-
-    public void setBackdropPath(String backdropPath) {
-        this.backdropPath = backdropPath;
-    }
-
-    public Boolean getAdult() {
-        return adult;
-    }
-
-    public void setAdult(Boolean adult) {
-        this.adult = adult;
-    }
-
-    public String getOverview() {
-        return overview;
-    }
-
-    public void setOverview(String overview) {
-        this.overview = overview;
-    }
-
-    public String getReleaseDate() {
-        return releaseDate;
-    }
-
-    public void setReleaseDate(String releaseDate) {
-        this.releaseDate = releaseDate;
-    }
-
-
-    @Override
-    public int describeContents() {
-        return 0;
-    }
-
-    @Override
-    public void writeToParcel(Parcel dest, int flags) {
-        if (voteCount == null) {
-            dest.writeByte((byte) 0);
-        } else {
-            dest.writeByte((byte) 1);
-            dest.writeInt(voteCount);
-        }
-        if (id == null) {
-            dest.writeByte((byte) 0);
-        } else {
-            dest.writeByte((byte) 1);
-            dest.writeInt(id);
-        }
-        dest.writeByte((byte) (video == null ? 0 : video ? 1 : 2));
-        if (voteAverage == null) {
-            dest.writeByte((byte) 0);
-        } else {
-            dest.writeByte((byte) 1);
-            dest.writeDouble(voteAverage);
-        }
-        dest.writeString(title);
-        if (popularity == null) {
-            dest.writeByte((byte) 0);
-        } else {
-            dest.writeByte((byte) 1);
-            dest.writeDouble(popularity);
-        }
-        dest.writeString(posterPath);
-        dest.writeString(originalLanguage);
-        dest.writeString(originalTitle);
-        dest.writeString(backdropPath);
-        dest.writeByte((byte) (adult == null ? 0 : adult ? 1 : 2));
-        dest.writeString(overview);
-        dest.writeString(releaseDate);
-    }
-}
