--- app/src/main/java/com/nanodegree/velaphi/popularmoviesstage1/models/TrailersResponse.java
@@ -1,33 +1,0 @@
-package com.nanodegree.velaphi.popularmoviesstage1.models;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
-
-public class TrailersResponse {
-
-    @SerializedName("id")
-    @Expose
-    private Integer id;
-
-    @SerializedName("results")
-    @Expose
-    private List<Trailer> trailers = null;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public List<Trailer> getTrailers() {
-        return trailers;
-    }
-
-    public void setResults(List<Trailer> trailers) {
-        this.trailers = trailers;
-    }
-}
