+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Popular Movies Stage 1</string>
+    <string name="app_name">Udacity Popular Movies</string>
     <string name="api_key">5028fde5e3652cfd9ac738403e7d0787</string>
     <string name="image_base_url">https://image.tmdb.org/t/p/w500</string>
     <string name="thumbnail_url">https://img.youtube.com/vi/%s/hqdefault.jpg"</string>
